refactor(header): replace Header.tsx with header/ orchestrator
Cuts over from the 402-line src/components/Header.tsx to the new
src/components/header/ folder: a thin orchestrator Header.tsx, an
index.ts barrel matching the shared-table precedent, and a smoke
test. AppShell.tsx now imports from @/components/header.

The leaf modules (icons, nav-config, useMenuController, menus,
DemoBanner) landed in earlier commits and are unchanged here.
